api/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
